+++ src/m365/teams/commands/channel/channel-member-remove.spec.ts
@@ -311,6 +311,43 @@ describe(commands.CHANNEL_MEMBER_REMOVE, () => {
         done(e);
       }
     });
+  });  
+
+  it('fails to get channel when channel does is not private', (done) => {
+    sinonUtil.restore(request.get);
+    sinon.stub(request, 'get').callsFake((opts) => {
+      if (opts.url === `https://graph.microsoft.com/v1.0/teams/${encodeURIComponent('00000000-0000-0000-0000-000000000000')}/channels?$filter=displayName eq '${encodeURIComponent('Other Channel')}'`) {
+        return Promise.resolve({
+          "value": [
+            {
+              "name": "Other Channel",
+              "membershipType": "standard"
+            }
+          ]
+        });
+      }
+
+      return Promise.reject('Invalid Request');
+    });
+
+    command.action(logger, {
+      options: {
+        teamId: '00000000-0000-0000-0000-000000000000',
+        channelName: 'Other Channel',
+        id: '00000',
+        confirm: true
+      }
+    }, (err?: any) => {
+      try {
+        assert.strictEqual(
+          JSON.stringify(err),
+          JSON.stringify(new CommandError(`The specified channel is not a private channel`)));
+        done();
+      }
+      catch (e) {
+        done(e);
+      }
+    });
   });
 
   it('correctly get channel id by channel name', (done) => {
@@ -319,7 +356,8 @@ describe(commands.CHANNEL_MEMBER_REMOVE, () => {
         return Promise.resolve({
           value: [
             {
-              "id": "19:00000000000000000000000000000000@thread.skype"
+              "id": "19:00000000000000000000000000000000@thread.skype",
+              "membershipType": "private"
             }
           ]
         });